REFACTOR existing test classes into `Feature` & `Unit` namespaces
